requirements.txt
flask
PyMuPDF
google-generativeai
dotenv
